set layout
Set controls in specific positions, deleted dataGridSpliter, put in button img Search
