+++ SimplyTotalCommander/MainWindow.xaml.cs
@@ -1,20 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿
 using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
 using System.Windows.Shapes;
-using System.Collections.ObjectModel;
-using System.IO;
-using ClassLibrary1;
+
 
 namespace SimplyTotalCommander
 {
@@ -26,9 +13,37 @@ namespace SimplyTotalCommander
         public MainWindow()
         {
             InitializeComponent();
+            Application.Current.MainWindow.WindowState = WindowState.Maximized;
         }
 
         private void WindowControl_Loaded(object sender, RoutedEventArgs e)
+        {
+
+        }
+        private void Button_Click_1(object sender, RoutedEventArgs e)
+        {
+
+        }
+        private void Button_Click_2(object sender, RoutedEventArgs e)
+        {
+
+        }
+        private void Button_Click_3(object sender, RoutedEventArgs e)
+        {
+
+        }
+
+        private void Button_Click_4(object sender, RoutedEventArgs e)
+        {
+
+        }
+
+        private void Button_Click_5(object sender, RoutedEventArgs e)
+        {
+
+        }
+
+        private void Button_Click_6(object sender, RoutedEventArgs e)
         {
 
         }
