+++ SimplyTotalCommander/WindowControl.xaml.cs
@@ -51,7 +51,6 @@ namespace SimplyTotalCommander
             DataGridOfFiles.ItemsSource = _fileList;
             SeachBox.ItemsSource = _fileList;
             ListOfDirectory.ItemsSource = listOfDirectoryNames;
-            DirectoriesTree.ItemsSource = listOfDirectoryNames;
         }
         private void DataGridOfFiles_SelectionChanged(object sender, SelectionChangedEventArgs e)
         {
@@ -61,36 +60,8 @@ namespace SimplyTotalCommander
         {
 
         }
-        private void Button_Click_1(object sender, RoutedEventArgs e)
-        {
-        //    fileReader.ProcessSubDirectories(NewPath.Text, listOfDirectoriesName, sender, new PropertyChangedEventArgs("listOfFiles"));
-        }
-
-        private void Button_Click_2(object sender, RoutedEventArgs e)
-        {
-
-        }
+ 
         private void DirectoryButton(object sender, RoutedEventArgs e)
-        {
-           // hiddenButton.Visiblity = Visibility.Collapsed; 
-        }
-
-        private void Button_Click_3(object sender, RoutedEventArgs e)
-        {
-
-        }
-
-        private void Button_Click_4(object sender, RoutedEventArgs e)
-        {
-
-        }
-
-        private void Button_Click_5(object sender, RoutedEventArgs e)
-        {
-
-        }
-
-        private void Button_Click_6(object sender, RoutedEventArgs e)
         {
 
         }
@@ -105,13 +76,11 @@ namespace SimplyTotalCommander
             if (item != null && item.IsSelected)
             {
                 string value = item.Content.ToString();
-                
                 foreach (var element in dir.GetDirectories(value,SearchOption.TopDirectoryOnly))
                 {
-
+                    string name = dir.Parent.ToString();
                     if (element.Name == value)
                     {
-
                         NewPath.Text = element.FullName;
                         Button_Click(sender,e);
                         //ICollectionView view = CollectionVie
